refactor(message): move utility functions to a dedicated utils module and update imports
