+++ packages/kit/src/core.ts
@@ -1,3 +1,3 @@
 export * from './message/core'
-export { combileDeltaData, normalizeToAsyncGenerator } from './message/core/utils'
 export * from './message/plugins'
+export { combileDeltaData, normalizeToAsyncGenerator } from './message/utils'
